Update introduction to vector databases
I started with updating the text to include a reference to the new Milvus implementation, but I added more context on vector databases and how they are used in the RAG pattern. 

This is the primary reason to use a vector database and I think we should mention it here.